# docs/mn/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Энэ жинээ нь Go хэл дээр Fiber framework-ийг ашиглаж EBarimt PosAPI-ийн REST API-г хурдан ба хэмнэлттэй бүтээсэн жишээ.",
        "title": "EBarimt PosAPI REST API Go ба Fiber ашигласан",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Munkh-Altai Chuluunbaatar",
            "email": "munkaltai@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/call": {
            "get": {
                "description": "Тодорхой нэртэй болон параметртэй функцыг дуудах.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ebarimt"
                ],
                "summary": "Тодорхой функцыг дуудах",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the function to call",
                        "name": "functionName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Parameters to pass to the function",
                        "name": "params",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/check": {
            "get": {
                "description": "API-ийн холболтыг шалгах мөн тохиргооны мэдээлл, сүлжээний статусыг аваh, өгөгдөлийн сантай харьцахад саад буй эсэхийг шалгаж болно",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ebarimt"
                ],
                "summary": "API-ийн холболтыг шалгах ба статусыг авах",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/posapi.CheckOutput"
                        }
                    }
                }
            }
        },
        "/info": {
            "get": {
                "description": "API-ийн талаарх мэдээллийг, жишээ нь хувилбар ба нэрээ авах.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ebarimt"
                ],
                "summary": "API-ийн мэдээллийг авах",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/posapi.InformationOutput"
                        }
                    }
                }
            }
        },
        "/organization/{register}": {
            "get": {
                "description": "Байгууллагын мэдээллийг регистрийн дугаараар авах",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ebarimt"
                ],
                "summary": "Байгууллагын мэдээллийг регистрийн дугаараар авах",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.OrganizationInfo"
                        }
                    }
                }
            }
        },
        "/put": {
            "post": {
                "description": "Өгөгдсөн мэдээллийг ашиглан системд нэхэмжлэхийг нэмэх.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ebarimt"
                ],
                "summary": "Системд нэхэмжлэхийг оруулах",
                "parameters": [
                    {
                        "description": "Bill data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/posapi.PutInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/posapi.PutOutput"
                        }
                    }
                }
            }
        },
        "/put-batch": {
            "post": {
                "description": "Өгөгдсөн мэдээллийг ашиглан системд нэхэмжлэхийг нэмэх.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ebarimt"
                ],
                "summary": "Системд багц баримтын нэхэмжлэхийг оруулах",
                "parameters": [
                    {
                        "description": "Bills data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/posapi.PutInputBatch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/posapi.PutOutput"
                        }
                    }
                }
            }
        },
        "/return": {
            "post": {
                "description": "Өгөгдсөн мэдээллийг ашиглан нэхэмжлэхийг буцаах.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ebarimt"
                ],
                "summary": "Нэхэмжлэхийг буцаах",
                "parameters": [
                    {
                        "description": "Return bill data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/posapi.BillInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/posapi.BillOutput"
                        }
                    }
                }
            }
        },
        "/send": {
            "get": {
                "description": "Баримтын мэдээллийг илгээх API.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ebarimt"
                ],
                "summary": "Баримтын мэдээллийг илгээх",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/posapi.DataOutput"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helper.OrganizationInfo": {
            "type": "object",
            "properties": {
                "citypayer": {
                    "type": "boolean"
                },
                "found": {
                    "type": "boolean"
                },
                "freeProject": {
                    "type": "boolean"
                },
                "lastReceiptDate": {},
                "name": {
                    "type": "string"
                },
                "receiptFound": {
                    "type": "boolean"
                },
                "vatpayer": {
                    "type": "boolean"
                },
                "vatpayerRegisteredDate": {
                    "type": "string"
                }
            }
        },
        "posapi.BankTransaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "approvalCode": {
                    "type": "string"
                },
                "bankId": {
                    "type": "string"
                },
                "rrn": {
                    "type": "string"
                },
                "terminalId": {
                    "type": "string"
                }
            }
        },
        "posapi.Bill": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "bankTransactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/posapi.BankTransaction"
                    }
                },
                "billIdSuffix": {
                    "type": "string"
                },
                "billType": {
                    "type": "string"
                },
                "branchNo": {
                    "type": "string"
                },
                "cashAmount": {
                    "type": "string"
                },
                "cityTax": {
                    "type": "string"
                },
                "customerNo": {
                    "type": "string"
                },
                "districtCode": {
                    "type": "string"
                },
                "internalId": {
                    "type": "string"
                },
                "invoiceId": {
                    "type": "string"
                },
                "nonCashAmount": {
                    "type": "string"
                },
                "posNo": {
                    "type": "string"
                },
                "registerNo": {
                    "type": "string"
                },
                "reportMonth": {
                    "type": "string"
                },
                "returnBillId": {
                    "type": "string"
                },
                "stocks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/posapi.Stock"
                    }
                },
                "taxType": {
                    "type": "string"
                },
                "vat": {
                    "type": "string"
                }
            }
        },
        "posapi.BillInput": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "returnBillId": {
                    "type": "string"
                }
            }
        },
        "posapi.BillOutput": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "posapi.CheckOutput": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string"
                        },
                        "success": {
                            "type": "boolean"
                        }
                    }
                },
                "database": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string"
                        },
                        "success": {
                            "type": "boolean"
                        }
                    }
                },
                "network": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string"
                        },
                        "success": {
                            "type": "boolean"
                        }
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "posapi.DataOutput": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "posapi.InformationOutput": {
            "type": "object",
            "properties": {
                "branchNo": {
                    "type": "string"
                },
                "dbDirPath": {
                    "type": "string"
                },
                "extraInfo": {
                    "type": "object",
                    "properties": {
                        "countBill": {
                            "type": "string"
                        },
                        "countLottery": {
                            "type": "integer"
                        },
                        "lastSentDate": {
                            "type": "string"
                        },
                        "posVersion": {
                            "type": "string"
                        }
                    }
                },
                "posId": {
                    "type": "string"
                },
                "registerNo": {
                    "type": "string"
                }
            }
        },
        "posapi.PutInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "bankTransactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/posapi.BankTransaction"
                    }
                },
                "billIdSuffix": {
                    "type": "string"
                },
                "billType": {
                    "type": "string"
                },
                "branchNo": {
                    "type": "string"
                },
                "cashAmount": {
                    "type": "string"
                },
                "cityTax": {
                    "type": "string"
                },
                "customerNo": {
                    "type": "string"
                },
                "districtCode": {
                    "type": "string"
                },
                "invoiceId": {
                    "type": "string"
                },
                "nonCashAmount": {
                    "type": "string"
                },
                "posNo": {
                    "type": "string"
                },
                "reportMonth": {
                    "type": "string"
                },
                "returnBillId": {
                    "type": "string"
                },
                "stocks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/posapi.Stock"
                    }
                },
                "taxType": {
                    "type": "string"
                },
                "vat": {
                    "type": "string"
                }
            }
        },
        "posapi.PutInputBatch": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "billIdSuffix": {
                    "type": "string"
                },
                "billType": {
                    "type": "string"
                },
                "bills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/posapi.Bill"
                    }
                },
                "group": {
                    "type": "boolean"
                },
                "posNo": {
                    "type": "string"
                },
                "vat": {
                    "type": "string"
                }
            }
        },
        "posapi.PutOutput": {
            "type": "object",
            "properties": {
                "billId": {
                    "type": "string"
                },
                "billType": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "errorCode": {
                    "type": "integer"
                },
                "internalCode": {
                    "type": "string"
                },
                "lottery": {
                    "type": "string"
                },
                "lotteryWarningMsg": {
                    "type": "string"
                },
                "macAddress": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "qrData": {
                    "type": "string"
                },
                "registerNo": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "posapi.Stock": {
            "type": "object",
            "properties": {
                "barCode": {
                    "type": "string"
                },
                "cityTax": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "measureUnit": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "qty": {
                    "type": "string"
                },
                "totalAmount": {
                    "type": "string"
                },
                "unitPrice": {
                    "type": "string"
                },
                "vat": {
                    "type": "string"
                }
            }
        }
    }
}